requirements.txt
accelerate==1.8.1
einops==0.8.1
ftfy==6.3.1
gguf==0.17.1
hf-transfer==0.1.9
huggingface-hub==0.33.1
imageio==2.37.0
loguru==0.7.3
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
peft==0.15.0
pillow==11.2.1
protobuf==6.31.1
pydash==8.0.5
pyyaml==6.0.2
requests==2.32.4
safetensors==0.5.3
tokenizers==0.21.2
tqdm==4.67.1
transformers==4.51.3
urllib3==2.5.0
google-cloud-storage
azure-storage-blob
boto3
tenacity
tiktoken
blobfile
sentencepiece
imageio-ffmpeg